[metrics] fix compilation on macOS
This is a follow-up to 926bca85827e597f2923652c3b3b017c7c071f3b.

Reviewed-on: http://gerrit.cloudera.org:8080/15055
Tested-by: Kudu Jenkins

+++ src/kudu/tablet/tablet_metrics.h
@@ -103,7 +103,7 @@ struct TabletMetrics {
   scoped_refptr<MeanGauge> average_diskrowset_height;
 
   // Static metrics.
-  scoped_refptr<AtomicGauge<uint64_t>> merged_entities_count_of_tablet;
+  scoped_refptr<AtomicGauge<size_t>> merged_entities_count_of_tablet;
 };
 
 } // namespace tablet
